[fix] 修改了赠送优惠券中的选择赠送用户功能

+++ src/main/java/com/example/cinema/bl/promotion/CouponService.java
@@ -17,7 +17,7 @@ public interface CouponService {
 
     ResponseVO issueCoupon(int couponId,int userId);
 
-    ResponseVO getUsersByConsume(int totalConsume);
+    ResponseVO getUsersByConsume(double totalConsume);
 
     ResponseVO presentCoupon2User(PresentForm presentForm);
 }
